extract nb top cell toolbar.
